+++ tf/client/make/help.py
@@ -1,14 +1,15 @@
 HELP = """
-text-fabric-make {dataset} serve {folder}
-text-fabric-make {dataset} ship
-text-fabric-make {dataset} {client} ship
-text-fabric-make {dataset} make [config] output
-text-fabric-make {dataset} {client} make [config] output
-text-fabric-make {dataset} {client} {command} [parameters]
+text-fabric-make {org}/{repo} serve {folder}
+text-fabric-make {org}/{repo} ship
+text-fabric-make {org}/{repo} {client} ship
+text-fabric-make {org}/{repo} make [config] output
+text-fabric-make {org}/{repo} {client} make [config] output
+text-fabric-make {org}/{repo} {client} {command} [parameters]
 
-dataset: a Text-Fabric dataset such as "nena", "bhsa"
+org/repo: a Text-Fabric dataset in a GitHub repo under org,
+          such as "etcbc/bhsa" and "CambridgeSemiticsLab/nena_tf"
 client: the name of a layered-search client as defined in the config.yaml
-        in annotation/app-«dataset»
+        in «org»/«repo»-search
 
 command:
 
@@ -21,10 +22,11 @@ v              : show current version of search-client code
 i              : increase version
 config         : build corpus config file
 corpus         : build corpus data file
-client         : build the layered-search client in the destination repo app-«dataset»
+client         : build the layered-search client in the destination «org»/«repo»-search
 clientdebug    : same as above but put it in debug mode
 debug on|off   : set debug flag of the client on or off
-publish        : publish the layered search client on the Github Pages of app-«dataset»
+publish        : publish the layered search client on the Github Pages of
+                 «org»/«repo»-«search»
 ship           : performs all build steps:
                  i, config, corpus, client, debug off, publish;
                  if {client} is given, ships this client,